Move display header and language file helpers into separate lib modules
